backend/requirements.txt
flask
flask-cors
flask-jwt-extended
pymongo
shap
joblib
pandas
numpy
python-dotenv
dnspython
